requirements.txt
# requirements.txt
fastapi
uvicorn
sqlalchemy
databases
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
 